Add CI / CD for the OS builders
This CI and CD pipeline will validate the packer files (we cannot run
a real build, as GitHub does not allow virt-in-virt).

It will also clone docker images for each major distro, then run the
provisioning playbook against each to ensure that they run correctly.
Currently this does nothing as the playbooks are empty.