BUG: Remove pip install python-debian
The python-debain package is a dependency for
ansible.builtin.deb8222_repository module. However, it is already
installed via apt either by default or by some other playbook. Dropping
this step as it was blocking ubuntu 24 images from building as you have
to install pip packages into a venv in noble.