MAINT: Ensure python3-debian is installed
python3-debian is required for the deb822 module but seems to be
installed by default on the Ubuntu 22 and 24 images. We should add it
here to make sure it is installed for future proofing. We should not use
pip to install it as Ubuntu noble encourages you to use venvs. Plus it
seems pip doesn't come installed by default on our Ubuntu 24 image.