BUG: Fix nubes-boot script failing due to shell differences
The ExecStart variable runs the nubes boot script with shell directly. This means it ignores the shebang at the top of the file which says to run with bash. By removing the call to shell and putting the direct file path to execstart it will use whichever interpreter the shebang states. This was an issue because rocky and ubuntu have different shell behaviours.