MAINT: Remove all users that are not cloud or os family
Editing this part of the tidy up to remove all users that are not either cloud, rocky or ubuntu. This is because when we build the VM with packer the fedID user add script adds the person who is building the machine to the VM even though we use the os distribution user to ssh in.