DOC: Update the README to reflect the new process
Adding more comprehensive instructions to how to use the image builders
for release and bug fixing.

Add python requirements for ansible version as Rocky 8 requires an old
version of ansible. Remove requirements for Ansible Galaxy as the only
collections listed are available in the default installation. Pin openstackcli version as it is most compatible for openstack yoga