LINT: Tidying the playbooks
Moving the become statement to each task rather than running the entire playbook as root. This is safer as any new task added will not be run as root by default. It is better practice.

Adding whitespaces to jinja templates and triple dashes to start files.

Removing ansible.cfg as the default role path is "roles" anyway. This was here because the playbooks were in a separate directory.