MAINT: Remove extra vars
We do not need the tidy image var as this is already declared on playbook level. If you do not want to tidy the image, don't include the role. And for our vm_baseline playbook we want the security baseline plus our fixes.

Remove provision this machine as it isn't working as there is no check. We also don't need it as localhost isn't implicitly matched to all. I.e you have to explicitly say run on localhost to do so.