ENH: Add role for image based fixes
Adds a new role for fixes to our images, e.g. options or files we need
to modify, add, or remove, from the original upstream distro.
This is typically because the generic defaults will cause problems, or
won't be optimal for OpenStack.

This is kept as its own role, as it's not required (like our VM
baseline) but is recommended, so people can choose if they'd like to
use these fixes. E.g. for troubleshooting or to eliminate them as a
potential cause of problems.