MAINT: Rename image playbook
Rename the `prepare_user_image.yml` to `configure_os_images.yml`. This is clearer for what the playbook does. It installs packages to the VM to comply with our security baseline and what fixes we need to work with openstack.

Remove provision_this_machine variable as it is not checked by the playbook and does not stop the play from running.

Flip tidy_image variable to always run by default. The only time we want this to not run is during the GitHub workflow as we cannot reboot localhost. Otherwise it should run as normal